pyproject.toml
[project]
name = "langchain-explained"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "groq>=0.30.0",
    "ipykernel>=6.30.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.28",
    "langchain-redis>=0.2.3",
    "langchainhub>=0.1.21",
    "openai>=1.98.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "youtube-search>=2.1.2",
    "youtube-transcript-api==1.1.0",
]
